+++ app/Controllers/Penjualan.php
@@ -5,6 +5,7 @@ namespace App\Controllers;
 use App\Models\PenjualanModel;
 use App\Models\ProdukModel;
 use App\Models\SatuanModel;
+use App\Models\Serverside_model;
 use CodeIgniter\I18n\Time;
 
 class Penjualan extends BaseController
@@ -13,6 +14,7 @@ class Penjualan extends BaseController
     protected $m_produk;
     protected $m_satuan;
     protected $validation;
+    protected $serversideModel;
 
     public function __construct()
     {
@@ -20,12 +22,12 @@ class Penjualan extends BaseController
         $this->m_produk = new ProdukModel();
         $this->m_satuan = new SatuanModel();
         $this->validation = \Config\Services::validation();
+        $this->serversideModel = new Serverside_model();
     }
     public function index()
     {
         $data['produk'] = $this->m_produk->getProduk();
         $data['satuan'] = $this->m_satuan->findAll();
-        $data['penjualan'] = $this->m_penjualan->getPenjualan();
         $data['title'] = 'Penjualan';
         $data['validation'] = $this->validation;
         return view('penjualan/v_penjualan', $data);
@@ -34,7 +36,7 @@ class Penjualan extends BaseController
     public function addPenjualan()
     {
         if (!$this->validate($this->validation->getRuleGroup('penjualan'))) {
-            session()->setFlashdata('info', 'error');
+            session()->setFlashdata('info', 'error_add');
             return redirect()->to('/penjualan')->withInput();
         }
         $produk = $this->request->getPost('produk');
@@ -77,8 +79,8 @@ class Penjualan extends BaseController
     public function editPenjualan()
     {
         if (!$this->validate($this->validation->getRuleGroup('produk'))) {
-            session()->setFlashdata('info', 'error');
-            return redirect()->to('/product')->withInput();
+            session()->setFlashdata('info', 'error_edit');
+            return redirect()->to('/penjualan')->withInput();
         }
         $id = $this->request->getPost('id');
         $nama_produk = $this->request->getPost('nama_produk');
@@ -87,25 +89,74 @@ class Penjualan extends BaseController
         $stok = $this->request->getpost('stok');
         $harga = $this->request->getPost('harga');
 
-        $this->m_produk->update($id, [
-            'nama_produk' => $nama_produk,
-            'id_cat_produk' => $category,
-            'id_satuan' => $satuan,
-            'harga' => $harga,
-            'stok' => $stok,
-        ]);
-        session()->setFlashdata('info', 'Data berhasil dirubah');
+        $query = $this->m_penjualan->find($id);
+        if ($query) {
+            $this->m_produk->update($id, [
+                'nama_produk' => $nama_produk,
+                'id_cat_produk' => $category,
+                'id_satuan' => $satuan,
+                'harga' => $harga,
+                'stok' => $stok,
+            ]);
+            session()->setFlashdata('info', 'Data berhasil dirubah');
+        } else {
+            session()->setFlashdata('info', 'error_edit');
+        }
+
         return redirect()->to('/product');
     }
 
     public function deletePenjualan($id)
     {
         if (!preg_match('/^[a-zA-Z0-9_]*$/', $id)) {
-            session()->setFlashdata('info', 'error');
+            session()->setFlashdata('info', 'error_delete');
             return redirect()->to('/penjualan');
         }
-        $this->m_penjualan->delete($id);
-        session()->setFlashdata('info', 'Data berhasil di hapus');
+
+        $query = $this->m_penjualan->find($id);
+        if ($query) {
+            $this->m_penjualan->delete($id);
+            session()->setFlashdata('info', 'Data berhasil di hapus');
+        } else {
+            session()->setFlashdata('info', 'error_delete');
+        }
+
         return redirect()->to('/penjualan');
+    }
+
+    public function listdata()
+    {
+        $column_order = array('nama_produk', 'nama_satuan', 'qty', 'total_jual', 'tanggal_jual', 'id');
+        $column_search = array('nama_produk', 'nama_satuan', 'qty', 'total_jual', 'tanggal_jual');
+        $order = array('tanggal_jual' => 'desc');
+        $list = $this->serversideModel->get_datatables('_data_penjualan', $column_order, $column_search, $order);
+        $data = array();
+        // $no = $this->request->getPost('start');
+        foreach ($list as $lt) {
+            $button_action = '<a href="#" class="btn btn-warning btn-circle edit" onclick="edit(\'' . $lt->id . '\')" data-toggle="modal" data-target="#penjualan_modal" title="Edit">
+                                <i class="fas fa-exclamation-triangle"></i>
+                              </a>
+                              <a href="#" class="btn btn-danger btn-circle delete" onclick="deleteData(\'_datpen\',\'' . $lt->id . '\')" title="Delete">
+                                <i class="fas fa-trash"></i>
+                              </a>';
+            $hasil_rupiah = "Rp" . number_format($lt->total_jual);
+
+            $row = array();
+            $row[] = $lt->nama_produk;
+            $row[] = $lt->nama_satuan;
+            $row[] = $lt->qty;
+            $row[] = $hasil_rupiah;
+            $row[] = $lt->tanggal_jual;
+            $row[] = $button_action;
+            $data[] = $row;
+        }
+        $output = array(
+            'draw' => $this->request->getPost('draw'),
+            'recordsTotal' => $this->serversideModel->count_all('_data_penjualan'),
+            'recordsFiltered' => $this->serversideModel->count_filtered('_data_penjualan', $column_order, $column_search, $order),
+            'data' => $data,
+        );
+
+        return json_encode($output);
     }
 }
